constants updated, regular progressive wave formulas implemented, tests to be created

+++ src/constants.jl
@@ -1,6 +1,6 @@
 # Constants file
 
 # Physics
-const g = 9.81 # gravitational acceleration (m/s²)
-const ρ = 1000 # water density (kg/m³)
-const tol = 10^-15 # error tolerance
+const g::Float64 = 9.81 # gravitational acceleration (m/s²)
+const ρ::Float64 = 1000 # water density (kg/m³)
+const tol::Float64 = 10^-15 # error tolerance
